Test mod using a dir, not a zip
Zip files get locked by Windows while the game is active and cannot be
touched, removed or updated. Directories are fine for some reason. Doing
testing via a directory therefore allows me to sync my game's copy of
the mod with the development version without needing to reboot the game.